refactor: split helicity module
